+++ RevitPlugin/Properties/AssemblyInfo.cs
@@ -1,0 +1,17 @@
+using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("Bimto3dPrint")]
+[assembly: AssemblyDescription("Revit plugin for building shell export.")]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("Bimto3dPrint")]
+[assembly: AssemblyProduct("Bimto3dPrint")]
+[assembly: AssemblyCopyright("Copyright © 2024")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
+
+[assembly: ComVisible(false)]
+[assembly: Guid("9b6f8d31-2a2d-4f9e-bfb7-82fbf031f1d5")]
+
+[assembly: AssemblyVersion("0.1.0.0")]
+[assembly: AssemblyFileVersion("0.1.0.0")]